backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.21.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
litellm>=0.6.0
firebase-admin>=6.2.0
python-multipart>=0.0.6
httpx>=0.24.0 